@@ -5,8 +5,8 @@ def GOOGLE_LOCATION_API_KEY = '"' + GoogleLocationAPIKey + '"' ?: '"No google lo
 def STRING = 'String'
 
 android {
-    compileSdkVersion 26
-    buildToolsVersion '26.0.2'
+    compileSdkVersion 27
+    buildToolsVersion '27.0.3'
 
     repositories {
         mavenCentral()
@@ -19,7 +19,7 @@ android {
     defaultConfig {
         applicationId "com.credolabs.justcredo"
         minSdkVersion 24
-        targetSdkVersion 26
+        targetSdkVersion 27
         versionCode 1
         versionName "1.0"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
@@ -68,34 +68,34 @@ dependencies {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
 
-    compile 'com.google.android.gms:play-services-auth:11.8.0'
-    compile 'com.google.android.gms:play-services-places:11.8.0'
-    compile 'com.google.android.gms:play-services-maps:11.8.0'
-    compile 'com.google.android.gms:play-services-location:11.8.0'
-    compile 'com.android.support:appcompat-v7:26.1.0'
+    compile 'com.google.android.gms:play-services-auth:12.0.0'
+    compile 'com.google.android.gms:play-services-places:12.0.0'
+    compile 'com.google.android.gms:play-services-maps:12.0.0'
+    compile 'com.google.android.gms:play-services-location:12.0.0'
+    compile 'com.android.support:appcompat-v7:27.1.0'
     compile 'com.android.support.constraint:constraint-layout:1.0.2'
     compile 'com.loopj.android:android-async-http:1.4.9'
-    compile 'com.android.support:design:26.1.0'
+    compile 'com.android.support:design:27.1.0'
     compile 'com.android.volley:volley:1.0.0'
     compile 'com.android.support:multidex:1.0.3'
     compile 'com.google.code.gson:gson:2.8.0'
     compile 'com.nineoldandroids:library:2.4.0'
     compile 'com.diegodobelo.expandingview:expanding-view:0.9.4'
     compile 'rebus:bottom-dialog:0.0.5'
-    compile 'com.github.bumptech.glide:glide:3.5.2'
     compile 'com.github.sujithkanna:smileyrating:1.6.5'
-    compile 'com.google.firebase:firebase-crash:11.8.0'
-    compile 'com.google.firebase:firebase-auth:11.8.0'
-    compile 'com.google.firebase:firebase-database:11.8.0'
-    compile 'com.google.firebase:firebase-messaging:11.8.0'
+    compile 'com.google.firebase:firebase-crash:12.0.0'
+    compile 'com.google.firebase:firebase-auth:12.0.0'
+    compile 'com.google.firebase:firebase-database:12.0.0'
+    compile 'com.google.firebase:firebase-messaging:12.0.0'
     compile 'com.firebase:firebase-client-android:2.5.2'
-    compile 'com.google.firebase:firebase-storage:11.8.0'
-    compile 'com.google.firebase:firebase-firestore:11.8.0'
+    compile 'com.google.firebase:firebase-storage:12.0.0'
+    compile 'com.google.firebase:firebase-firestore:12.0.0'
     compile 'com.firebaseui:firebase-ui:1.2.0'
+    compile 'com.firebaseui:firebase-ui-firestore:3.2.2'
     compile 'com.theartofdev.edmodo:android-image-cropper:2.4.+'
     compile 'com.github.aakira:expandable-layout:1.4.2@aar'
     compile 'com.github.myinnos:AwesomeImagePicker:1.0.2'
-    compile 'com.google.android.gms:play-services-ads:11.8.0'
+    compile 'com.google.android.gms:play-services-ads:12.0.0'
     testCompile 'junit:junit:4.12'
     compile "org.jetbrains.kotlin:kotlin-stdlib-jre7:$kotlin_version"
 }
