@@ -4,6 +4,7 @@ import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
+import android.support.annotation.NonNull;
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -72,7 +73,7 @@ public class CategoryGridFragment extends Fragment {
     }
 
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.fragment_category_grid, container, false);
         TextView textName = (TextView) view.findViewById(R.id.category_name);
@@ -154,9 +155,9 @@ public class CategoryGridFragment extends Fragment {
             divider.setVisibility(View.GONE);
         }
         if (adapter == null) {
-            adapter = new CategoryGridAdapter(getActivity(), schoolArrayList);
+            //adapter = new CategoryGridAdapter(getActivity(), schoolArrayList);
 
-            grid.setAdapter(adapter);
+            //grid.setAdapter(adapter);
         } else {
             adapter.notifyDataSetChanged();
         }
