@@ -1,5 +1,6 @@
 package com.credolabs.justcredo.autocomplete;
 
+import android.annotation.SuppressLint;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -96,7 +97,7 @@ public class CustomPlaceAutoCompleteFragment extends PlaceAutocompleteFragment {
         int var1 = -1;
 
         try {
-            Intent var2 = (new PlaceAutocomplete.IntentBuilder(2)).setBoundsBias(this.zzaRk).setFilter(this.zzaRl).zzim(this.editSearch.getText().toString()).zzdz(1).build(this.getActivity());
+            @SuppressLint("RestrictedApi") Intent var2 = (new PlaceAutocomplete.IntentBuilder(2)).setBoundsBias(this.zzaRk).setFilter(this.zzaRl).zza(this.editSearch.getText().toString()).zza(1).build(this.getActivity());
             this.startActivityForResult(var2, 1);
         } catch (GooglePlayServicesRepairableException var3) {
             var1 = var3.getConnectionStatusCode();
