@@ -25,10 +25,10 @@ public class CustomeToastFragment {
 
         // inflate the layout over view
         View layout = inflater.inflate(R.layout.custom_toast,
-                (ViewGroup) view.findViewById(R.id.toast_root));
+                view.findViewById(R.id.toast_root));
 
         // Get TextView id and set error
-        TextView text = (TextView) layout.findViewById(R.id.toast_error);
+        TextView text = layout.findViewById(R.id.toast_error);
         text.setText(error);
 
         Toast toast = new Toast(context);// Get Toast Context
